Creates data migration tooling for v1 to v2 forms
Adds files and tooling necessary to perform a data migration from the v1 form structure to the v2 form structure.

This includes:
-A field mapper to assist in mapping v1 fields to v2 fields
-A form converter to perform the data transformation
-A template for the v2 form structure
-Authentication utilities

The goal of these files is to generate an output that can be compared to the actual v2 event and identify any missing or ambigious mappings. This should allow for auto generating translations from v1 client.csv and identify differences between the two forms.